2.0
Added bower package management for angular, bootstrap, and font awesome. Added instructions for read me. Edited AddSrv functionality into ListSrv instead because the listen/broadcast isn't efficient when dealing with the playlist. The ListSrv acts as the playlist as well as the list in general; it manages the play index and can grab the previous and next video. Added editing in place in the list instead of hiding the list entirely. Added buttons listeners for previous and next so that the user can control with fingahs. Added loading animation of 3 seconds and hid the video when going between videos.