Components/Shared/Routing
Added in ng-route. Moved current 'components' into 'directives' and made actual 'components' which are now the two pages of the app. The two pages of the apps templates live in index.html since this is running locally and not on a host. The components reuse the shared directives and providers which maintain states when navigating to and from the pages.